+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `customPixelEventId` to `modal-layout` block.
+
 ## [0.10.0] - 2021-04-22
 
 ### Added